BOSH support for TSUNG (TECH-1044).
To use BOSH, you need to change the server config of tsung, instead of

<server .... type="tcp"  port="5222"></server>
you must use:
<server .... type="bosh" port="5280"></server>

Note that BOSH uses xmpp 1.0, one must use SASL to authenticate.
In the generated reports, a new graph is included, showing
the # of HTTP connections and requests made. # of requests should
be much bigger than # of connections, as we are using persistent
connections.

Reported network traffic includes overhead by BOSH/HTTP protocol for POST
request it *DO NOT* include the HTTP overhead for responses.

TODO:
	normal HTTPS sessions are broken, I need to add ts_ssl implementing
        the same API than ts_tcp, ts_bosh.