Privacy lists rudimentary support
privacy:get_names gets the list of all names of privacy lists stored by
the server for a given user.

privacy:set_active sets a list with a predefined name as active. The list
name is determined from the JID, e.g. if the user's JID
is "john@average.com" then the list name is "john@average.com_list".
One should take care of properly seeding the server database in order to
ensure that such a list exists.