Record extra headers in the HTTP proxy recorder
Many web apps change the response based on the Accept: header in order to support both AJAX and traditional HTML endpoints on the same URLs (and in the past it was common also to look for XMLHttpRequest in the x-requested-with header, though that is used less and less these days).  And the Cache-Control: and Pragma: headers often affect whether requests go through to the app tier.