Allow path to HTML templates to be specified from absolute pathname.
If you set the "template_path" environment variable for the
ts_controller app, tsung will use that path instead of computing it
from the location of the controller beam file.

The original behavior is preserved by using the 'beam_relative' atom
as the value of template_path, which is the default in
ts_controller.app.